Restructuring modules for models
